maint: Reorganize and rename subfolders in `/analyzers` for consistency (#207)
* chore: rename folder hashtags --> hashtags_base

* chore: move *_base and *_web into hashtags folder

* chore: move test_data one level up

* chore: update all hashtags imports

* chore: remove temporal.ipynb

* chore: rename temporal --> temporal_base

* chore: rename temporal_barplot --> temporal_web for consistency

* move _base and _web into /analyzers/temporal

* chore: add temporal/__init__.py

* chore: update temporal imports

* chore: rename test_hashtags_analyzers.py

* chore: remove unused and commented code
