--- analyzers/hashtags/__init__.py
@@ -1,6 +1,0 @@
-from analyzer_interface import AnalyzerDeclaration
-
-from .interface import interface
-from .main import main
-
-hashtags = AnalyzerDeclaration(interface=interface, main=main, is_distributed=True)
