+++ analyzers/hashtags/hashtags_base/interface.py
