--- analyzers/temporal/__init__.py
@@ -1,6 +1,0 @@
-from analyzer_interface import AnalyzerDeclaration
-
-from .interface import interface
-from .main import main
-
-temporal = AnalyzerDeclaration(interface=interface, main=main)
