+++ analyzers/temporal/temporal_base/__init__.py
@@ -1,0 +1,6 @@
+from analyzer_interface import AnalyzerDeclaration
+
+from .interface import interface
+from .main import main
+
+temporal = AnalyzerDeclaration(interface=interface, main=main)
