+++ analyzers/temporal/temporal_web/__init__.py
@@ -3,6 +3,6 @@ from analyzer_interface import WebPresenterDeclaration
 from .factory import factory
 from .interface import interface
 
-temporal_barplot = WebPresenterDeclaration(
+temporal_web = WebPresenterDeclaration(
     interface=interface, factory=factory, name=__name__, shiny=False
 )
